requirements.txt
yt-dlp>=2023.12.30
pyyaml>=6.0.1
requests>=2.31.0
python-dotenv>=1.0.0